Con Configuracion

+++ src/tasks/tasks.service.ts
@@ -17,12 +17,16 @@ export class TasksService {
     ) {}
 
     
-    async getTasks(filterTask: GetTasksFilterDto): Promise<Task[]> {
-        return await this.taskRepository.getTasks(filterTask);
+    async getTasks(filterTask: GetTasksFilterDto, user: User): Promise<Task[]> {
+        return await this.taskRepository.getTasks(filterTask, user);
     }    
 
-    async getTaskById(id: number): Promise<Task> {
-        const found =  await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User
+        ): Promise<Task> {
+        
+            const found =  await this.taskRepository.findOne({where: {id, userId: user.id}});
 
         if (!found) {
             throw new NotFoundException(`Task id [${id}] Not Found`);
@@ -37,16 +41,28 @@ export class TasksService {
       return await this.taskRepository.createTask( createTaskDto, user);
     }
    
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(
+        id: number,
+        user: User
+        
+        ): Promise<void> {
+
+        const result = await this.taskRepository.delete({id, userId: user.id});
+
         if (result.affected === 0) {
             throw new NotFoundException(`Task id [${id}] Not Found`);
         }        
         return ;
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number, 
+        status: TaskStatus,
+        user: User
+    
+        ): Promise<Task> {
+    
+        const task = await this.getTaskById(id, user);
         task.status = status;
         return await this.taskRepository.save(task);
     }
